requirements.txt
colorama
shodan
python-dotenv

